fix(currentElectionStore): move barChart values to vue template
